Back to Stage_6
Found that there are logs in GitHub and NetBeans. As GitHub is targeting
the GitHub folder, which the GitHub log file aiming to, the source
folder may not change anything if there are any changes in the GitHub
folder. Because the log file in NetBeans consider the newest updates
only, it will make no updates even if the GitHub folder has any changes
(when doing the Revert). Then the conflicts happen. Suggestion: target
the source folder as the GitHub folder. After coding, export the souce
files into somewhere else, as the final products.